efficiency function done

+++ MNISTread.h
@@ -4,7 +4,7 @@
 
 
 int ReverseInt(int i);
-void read_Mnist_Image(std::string filename, std::vector<std::vector<double> > &vec);
-void read_Mnist_Label(std::string filename, std::vector<double> &vec);
+void read_Mnist_Image(std::string filename, std::vector<std::vector<unsigned int> > &vec);
+void read_Mnist_Label(std::string filename, std::vector<unsigned int> &vec);
